chore: Enhance sentry events & tracing (#547)
This PR changes the way events are submitted to Sentry, so we will have
the following structure of message
* Title: Error/Warning/panic
* Message: Actual message logged in trace

All the other required debugging information will be contained in
tracing fields.

To prevent receiving a lot of (not always readable)text in messages,
this PR also changes the way we are logging warnings and errors:
* Actual message contains only short description of what happened
* All the non-constant values are moved to tracing fields, *except*
components name - so the message will clearly state which component is
failed

Instead of sending `POD_NAMESPACE` as environment tag, send it as
separate namespace tag.