feat: revm consistency checker (#525)
Adding a separate pipeline item for running the revm execution
environment to compare state results. The zksync-revm uses a Database
with access to a custom state handler. All state changes across one
block are accumulated in an in-memory (cache) database, and for each new
block, the data is loaded from the server state handler. This means that
each block comparison is done separately, and if a mismatch happens in
block X, the block X + 1 comparison won’t be affected.

You can find the zksync-revm here:
- https://github.com/matter-labs/zksync-os-revm/pull/2

---------

Co-authored-by: romanbrodetski <rb@matterlabs.dev>