feat: Protocol upgrade support for provers (#577)
RATIONALE: Allow protocol upgrades to happen, without provers trying to
pick up "wrong batches" (read as batches that it cannot prove).

The changes force server <-> prover interaction to be VK based.

The new VK concept is a 1o1 mapping between `ExecutionVersion` and
`VerificationKeyHash`. Provers by default will provide a list of VKs
they support and server will provide jobs that correspond to such
versions.

This is a "breaking" change, as in, communication between the 2
components changes. The implementation is backwards compatible (legacy
API & v1 API).

The plan is to remove legacy API by end of Q4 2025. Until then, the code
is a bit "dirty" as we need to support 2 versions.

1 implication/draw-back of current implementation: We won't be able to
generate proof input on ExecutionVersion A and prove it with
ExecutionVersion B (that implies deprecating & removing the
execution_version on proofs).