feat: get rid of batch rescheduling (preparation to get rid of BatchStorage) (#587)
## WHAT

At node startup, it's common to have a number of blocks that are already
committed - but not executed or proved yet.

### Before this change,
we were handling this case explicitely - determining the list of such
batches, loading them from proof storage, and prepending the organic
input to certain components with this vector. For example, we were
sending the unproved batches to SnarkJobManager step so that it
processes them BEFORE it starts to get its organic input from
L1Committer.
Even if we replayed such blocks in sequencer, Batcher component would
simply drop them and not pass forward - **batcher was only ever
producing new batches** - the ones that were never committed to L1. All
the components in Batcher Subsystem were only working with such "new"
batches - most notably, l1 senders always expected to get a batch that's
not processed on L1 yet (e.g. L1ProveSender expected that its first
input batch is precisely the first unproved on L1 batch)

### After this change, 
all the components are capable of replaying arbitrary batches:
* **Batcher** recreates and asserts batches when handling blocks that
were already committed. As soon as it receives a block from a "new" (not
yet committed) batch - it switches to creation mode
* **L1 Senders** now support the following commands:
```
/// Batches that are already committed/proved may also go through the pipeline.
/// For such batches, a Passthrough variant is generated.
/// For batches that have to be processed on L1, a SendToL1 variant is used.
pub enum L1SenderCommand<Command: SendToL1> {
    SendToL1(Command),
    Passthrough(Box<BatchEnvelope<FriProof>>),
}
```
* **SnarkProverGenerator** only (re)generates proofs for batches that
are not proved on L1 - using the corresponding variant of the enum
above.

With all these changes, unproved/unexecuted on l1 batches are finalized
naturally because corresponding blocks pass through the pipeline in
full.

Additionally, this PR streamlines the pipeline starting block
determination.

## WHY

1. To get rid of proof storage (there will be no place to load the
batches to rescheulde from)
2. Being able to (re)process batches in all the components is useful for
operations/debug - e.g. we may want to check consistency, regenerate a
SNARK proof from freshly computed FRIs, etc.

## CAVEATS

* we now regenerate fri proofs for all blocks that are committed but not
executed. We can optimize this by storing generated FRI proofs on disc.
The key difference with the current approach is that this is just a
cache - so if we were to rotate the leader, the new main node would
seamlessly regenerate them.
* initially I expected the changeset to be smaller. But in the end I had
to introduce this `L1SenderCommand::Passthrough` command - and also in
some components we now have two phases/cases - when we process the input
anew, or reprocess it. In my opinion, it's worth it - but alternative
suggestions or concerns are welcome.

---------

Co-authored-by: perekopskiy <53865202+perekopskiy@users.noreply.github.com>